expand s3fs testing
