+++ minimalkv/net/boto3store.py
@@ -266,135 +266,9 @@ class Boto3Store(KeyValueStore, UrlMixin, CopyMixin):  # noqa D
             and self.bucket.name == other.bucket.name
             and self.bucket.meta.client.meta.endpoint_url
             == other.bucket.meta.client.meta.endpoint_url
-            and self.prefix == other.prefix
+            and self._object_prefix == other._object_prefix
             and self.url_valid_time == other.url_valid_time
             and self.reduced_redundancy == other.reduced_redundancy
             and self.public == other.public
             and self.metadata == other.metadata
         )
-
-    @classmethod
-    def from_url(cls, url: str) -> "Boto3Store":
-        """
-        Create a ``Boto3Store`` from a URL.
-
-        URl format:
-        ``s3://access_key_id:secret_access_key@endpoint/bucket[?<query_args>]``
-
-        **Positional arguments**:
-
-        ``access_key_id``: The access key ID of the S3 user.
-
-        ``secret_access_key``: The secret access key of the S3 user.
-
-        ``endpoint``: The endpoint of the S3 service. Leave empty for standard AWS.
-
-        ``bucket``: The name of the bucket.
-
-        **Query arguments**:
-
-        ``force_bucket_suffix`` (default: ``True``): If set, it is ensured that
-         the bucket name ends with ``-<access_key>``
-         by appending this string if necessary.
-         If ``False``, the bucket name is used as-is.
-
-        ``create_if_missing`` (default: ``True`` ): If set, creates the bucket if it does not exist;
-         otherwise, try to retrieve the bucket and fail with an ``IOError``.
-
-        **Notes**:
-
-        If the scheme is ``hs3``, an ``HBoto3Store`` is returned which allows ``/`` in key names.
-
-        Parameters
-        ----------
-        url
-            URL to create store from.
-
-        Returns
-        -------
-        store
-            Boto3Store created from URL.
-        """
-        from minimalkv import get_store_from_url
-
-        store = get_store_from_url(url, store_cls=cls)
-        if not isinstance(store, cls):
-            raise ValueError(f"Expected {cls}, got {type(store)}")
-        return store
-
-    @classmethod
-    def from_parsed_url(
-        cls, parsed_url: SplitResult, query: Dict[str, str]
-    ) -> "Boto3Store":  # noqa D
-        """
-        Build a Boto3Store from a parsed URL.
-
-        See :func:`from_url` for details on the expected format of the URL.
-
-        Parameters
-        ----------
-        parsed_url: SplitResult
-            The parsed URL.
-        query: Dict[str, str]
-            Query parameters from the URL.
-
-        Returns
-        -------
-        store : Boto3Store
-            The created Boto3Store.
-        """
-
-        url_access_key_id = _get_username(parsed_url)
-        url_secret_access_key = _get_password(parsed_url)
-        boto3_params = {
-            "aws_access_key_id": url_access_key_id,
-            "aws_secret_access_key": url_secret_access_key,
-        }
-        host = parsed_url.gethost()
-        port = parsed_url.getport()
-
-        is_secure = query.get("is_secure", "true").lower() == "true"
-        endpoint_scheme = "https" if is_secure else "http"
-
-        if host is None:
-            endpoint_url = None
-        elif port is None:
-            endpoint_url = f"{endpoint_scheme}://{host}"
-        else:
-            endpoint_url = f"{endpoint_scheme}://{host}:{port}"
-
-        boto3_params["endpoint_url"] = endpoint_url
-
-        # Remove Nones from client_params
-        boto3_params = {k: v for k, v in boto3_params.items() if v is not None}
-
-        bucket_name = parsed_url.getpath().lstrip("/")
-
-        # We only create a reference to the bucket.
-        # The bucket currently has to be created outside minimalkv.
-        # When the Boto3Store is reimplemented using s3fs,
-        # we will create the bucket in the `create_filesystem` method.
-        # Or we can just pass `mkdir_prefix` to the FSSpecStore constructor.
-        resource = boto3.resource("s3", **boto3_params)
-
-        force_bucket_suffix = query.get("force_bucket_suffix", "true").lower() == "true"
-        if force_bucket_suffix:
-            # Try to find access key in env
-            if url_access_key_id is None:
-                access_key_id = os.environ.get("AWS_ACCESS_KEY_ID")
-            else:
-                access_key_id = url_access_key_id
-
-            if access_key_id is None:
-                raise ValueError(
-                    "Cannot find access key in URL or environment variable AWS_ACCESS_KEY_ID"
-                )
-
-            if not bucket_name.lower().endswith("-" + access_key_id.lower()):
-                bucket_name += "-" + access_key_id.lower()
-
-        bucket = resource.Bucket(bucket_name)
-
-        create_if_missing = query.get("create_if_missing", "true").lower() == "true"
-
-        return cls(bucket, create_if_missing=create_if_missing)
